+++ RestAPI/RestAPI/Models/ProductInSiteModel.cs
@@ -7,7 +7,7 @@ namespace RestAPI.Models
         public string CompID { get; set; }
         public string SiteID { get; set; }
         public string ProdID { get; set; }
-        public decimal? Quantity { get; set; }
+        public double? Quantity { get; set; }
         public string UOM { get; set; }
         public IEnumerable<OrderDetailModel> OrderDetails { get; set; }
